leasehttp: move lease/http.go to its own pkg
